Move bibtex on main back in multibib if statement.
This basically undo's commit acc72788dd8cde419aa3dd4e7020076188aaf8d9.

Apparently, there IS harm in running bibtex on an aux file with no cites
in it.  Bibtex will give you an error in this case.  It is also unclear
whether it is even possible to use the standard cites when using bibtex
or if you ever want to.  If you do, there is still the fallback of
adding the main target to the MULTIBIB_NEWCITES command, if you know
enough to do so.